file/{download,crawl}: rationalize use of file.FS etc (#114)
Modify the downloader API to return the contents of the download file and their metadata rather than writing them to file.WriteFS. file.WriteFS is removed. file.FS is modified to extend fs.FS and for it's Open method to be renamed as OpenCtx to signify its acceptance of a context parameter. This is useful/necessary for implementing file.FS over systems such as AWS S3 which expect a context for which cancelation is a necessary feature.