file,cmdutil: add FSReadFile, FSOpen that extract an fs.ReadFileFS from the context and use that to read a file (#180)
FSReadFile, FSOpen extract an fs.ReadFileFS from the context and use that to read the requested file if present, otherwise they use os.Open and os.ReadFile. This allows for files to be located locally, remotely or embedded and be accessed via the same API. This is primarily intended for small, configuration and similar files.