file/filewalk: remove async status reports, improve context cancelation (#224)
The introduction of the 'state' type to the Handler makes it straightforward for the client to track timing information and hence the status callback can be removed and the walker simplified. In addition, the checks for context cancelation are relocated to prevent new traverses from being initiated after the context is being canceled to speed up returning from Walk when the context is canceled.