aws/s3fs: stat validates a prefix by calling ListPrefix on it (#288)
Improve handling of Stat for prefixes vs objects. If a path ends in / (or whatever delimiter is in use) then try it as a prefix first and then as an object. If path does not end in / then reverse the order, try head first and then as a prefix. The test for a valid prefix is simply to list the prefix and see if it has any contents.